+++ src/translations/en/notes.php
@@ -8,9 +8,19 @@
 
 return [
 	'Notes' => 'Notes',
-	'Allow Deleting' => 'Allow Deleting',
-	'Allow users to delete notes' => 'Allow users to delete notes',
 	'You must save the element before you can add notes!' => 'You must save the element before you can add notes!',
 	'Add a note' => 'Add a note',
 	'Add Note' => 'Add Note',
+	'Add notes' => 'Add notes',
+	'Delete own notes' => 'Delete own notes',
+	'Delete all notes' => 'Delete all notes',
+
+	'All users' => 'All users',
+	'User permission' => 'User permission',
+	'Disabled' => 'Disabled',
+
+	'Adding' => 'Adding',
+	'Deleting' => 'Deleting',
+	'Manage who can add notes' => 'Manage who can add notes',
+	'Manage who can delete notes' => 'Manage who can delete notes',
 ];